pyproject.toml
[project]
name = "msprogramcontrol"
version = "1.2.0"
description = "GUI interface for control and upload programs to the Lego Robot Inventor (51515) hub and the SPIKE Prime (45678) hub."
authors = [
    {name = "Tibor", email = ""},
]
dependencies = [
    "pyserial>=3.5",
    "mpy-cross-v5>=1.0.1",
    "async-tkinter-loop>=0.9.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.setuptools.package-data]
"msprogramcontrol.assets" = ["*.png"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
